requirements.txt
# Python 3.9
django==3.1
django-debug-toolbar==3.2
pillow==8.0
tzlocal==2.1
babel==2.9
pytz~=2020.5
selenium~=3.141.0
mysqlclient
requests==2.25.1
djangorestframework==3.12.2